# docs/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "CLARK Gateway",
    "version": "3.7.6",
    "description": "The API gateway for all of the CLARK microservices",
    "license": { "name": "ISC", "url": "https://www.isc.org/licenses/" },
    "contact": {
      "name": "SecurEd Inc.",
      "url": "https://secured.team/",
      "email": "info@secured.team"
    },
    "x-logo": {
      "url": "https://clark.center/assets/images/logo.png",
      "altText": "CLARK Logo"
    },
    "x-tagGroups": [
      { "name": "Library Service", "tags": ["Library Service"] },
      { "name": "Notification Service", "tags": ["Notification Service"] },
      { "name": "Feature Service", "tags": ["Feature Service"] },
      { "name": "Rating Service", "tags": ["Rating Service"] },
      { "name": "Outcome Service", "tags": ["Outcome Service"] },
      {
        "name": "Learning Object Service",
        "tags": ["Learning Object Service"]
      },
      { "name": "User Service", "tags": ["User Service"] },
      { "name": "Utility Service", "tags": ["Utility Service"] },
      {
        "name": "Standard Guideline Service",
        "tags": ["Standard Guideline Service"]
      }
    ]
  },
  "servers": [
    { "url": "http://localhost:3001", "description": "Development" },
    { "url": "https://api-gateway.clark.center", "description": "Production" }
  ],
  "paths": {
...
    "/users/{username}/notifications": {
      "get": {
        "description": "Gets a user's update notifications",
        "tags": ["Notification Service"],
        "parameters": [
          {
            "in": "path",
            "name": "username",
            "schema": { "type": "string" },
            "required": true,
            "description": "The username of the user"
          },
          {
            "in": "query",
            "name": "page",
            "schema": { "type": "number" },
            "required": true,
            "description": "The current page"
          },
          {
            "in": "query",
            "name": "limit",
            "schema": { "type": "number" },
            "required": true,
            "description": "The notification limit to send back in the request"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "lastPage": {
                      "type": "number",
                      "example": 1,
                      "required": true
                    },
                    "notifications": {
                      "type": "array",
                      "items": { "$ref": "#/components/schemas/Notification" },
                      "required": true
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "BAD CONTENT - Page and limit query parameters must be greater then zero"
          },
          "401": { "description": "UNAUTHENTICATED - User not logged in" },
          "403": {
            "description": "UNAUTHORIZED - User is trying to access another user's notifications"
          },
          "404": { "description": "NOT FOUND - User not found" }
        }
      }
    },
    "/users/{username}/notifications/{id}": {
      "delete": {
        "description": "Deletes a notification for a user",
        "tags": ["Notification Service"],
        "parameters": [
          {
            "in": "path",
            "name": "username",
            "schema": { "type": "string" },
            "required": true,
            "description": "The username of the user"
          },
          {
            "in": "path",
            "name": "id",
            "schema": { "type": "string" },
            "required": true,
            "description": "The id of the notification"
          }
        ],
        "responses": {
          "204": { "description": "NO CONTENT" },
          "401": { "description": "UNAUTHENTICATED - User not logged in" },
          "403": {
            "description": "UNAUTHORIZED - User is trying to delete another user's notifications"
          },
          "404": { "description": "NOT FOUND - User or notification not found" }
        }
      }
    },
    "/users/{username}/learning-objects/{cuid}/notifications": {
      "post": {
        "description": "Create a notification for a object",
        "tags": ["Notification Service"],
        "parameters": [
          {
            "in": "path",
            "name": "username",
            "schema": { "type": "string" },
            "required": true,
            "description": "The username of the author of the learning object"
          },
          {
            "in": "path",
            "name": "cuid",
            "schema": { "type": "string" },
            "required": true,
            "description": "The cuid of the object to make the notification for"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/NotificationPayload" }
            }
          }
        },
        "responses": {
          "201": { "description": "CREATED" },
          "400": { "description": "BAD REQUEST - Request body invalid" },
          "401": { "description": "UNAUTHENTICATED - User is not logged in" },
          "403": {
            "description": "UNAUTHORIZED - User is not admin or editor"
          },
          "404": { "description": "NOT FOUND - Learning object not found" }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Notification": {
        "properties": {
          "id": {
            "type": "string",
            "description": "The id of the notification",
            "example": 0,
            "required": true
          },
          "text": {
            "type": "string",
            "description": "The description of the notification",
            "example": "This is an example",
            "required": true
          },
          "timestamp": {
            "type": "string",
            "description": "The timestamp for when the notification was created",
            "example": 0,
            "required": true
          },
          "link": {
            "type": "string",
            "description": "The link to the object",
            "example": "https://clark.center/details/jdoe1/00000000-1111-2222-3333-444444444444",
            "required": true
          },
          "attributes": {
            "type": "object",
            "required": true,
            "additionalProperties": { "type": "string" }
          }
        }
      },
      "NotificationPayload": {
        "properties": {
          "username": {
            "type": "string",
            "description": "The requester's username",
            "example": "jdoe1",
            "required": true
          },
          "author": {
            "type": "string",
            "description": "The learning object author's username",
            "example": "jdoe1",
            "required": true
          },
          "learningObjectName": {
            "type": "string",
            "description": "The name of the learning object updated",
            "example": "This is a learning object",
            "required": true
          },
          "version": {
            "type": "number",
            "description": "The version number of the object",
            "example": 0,
            "required": true
          },
          "cuid": {
            "type": "string",
            "required": true,
            "description": "The cuid of the learning object",
            "example": "aaaaaaaa-bbbb-cccc-dddd-eeeeeeeeeeee"
          },
          "authorID": {
            "type": "string",
            "required": true,
            "description": "The learning object author's id",
            "example": 0
          }
        }
      },
...
    }
  },
...
}